// src/api/generated/model/cartItemResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Ecommerce Copa API
 * API backend do Ecommerce Copa com autenticação por sessão, catálogo público e CRUD administrativo.
 * OpenAPI spec version: 1.0.0
 */
import type { CartItemResponseAttributes } from "./cartItemResponseAttributes.ts";

export interface CartItemResponse {
	skuId?: string;
	productId?: string;
	title?: string;
	unitPrice?: number;
	photo?: string;
	stock?: number;
	amount?: number;
	subtotal?: number;
	attributes?: CartItemResponseAttributes;
}